+++ src/main/java/xyz/tomorrowlearncamp/bookking/domain/user/auth/service/AuthService.java
@@ -15,8 +15,8 @@ import xyz.tomorrowlearncamp.bookking.domain.user.auth.dto.SignupRequest;
 import xyz.tomorrowlearncamp.bookking.domain.user.auth.dto.SignupResponse;
 import xyz.tomorrowlearncamp.bookking.domain.user.auth.entity.RefreshToken;
 import xyz.tomorrowlearncamp.bookking.domain.user.auth.repository.RefreshTokenRepository;
-import xyz.tomorrowlearncamp.bookking.domain.user.dto.LoginRequest;
-import xyz.tomorrowlearncamp.bookking.domain.user.dto.LoginResponse;
+import xyz.tomorrowlearncamp.bookking.domain.user.dto.request.LoginRequest;
+import xyz.tomorrowlearncamp.bookking.domain.user.dto.response.LoginResponse;
 import xyz.tomorrowlearncamp.bookking.domain.user.entity.User;
 import xyz.tomorrowlearncamp.bookking.domain.user.repository.UserRepository;
 import org.springframework.http.HttpHeaders;
