+++ src/main/java/xyz/tomorrowlearncamp/bookking/domain/user/service/UserService.java
@@ -5,9 +5,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
-import xyz.tomorrowlearncamp.bookking.domain.user.dto.UpdateUserRequest;
-import xyz.tomorrowlearncamp.bookking.domain.user.dto.UpdateUserRoleRequest;
-import xyz.tomorrowlearncamp.bookking.domain.user.dto.UserResponse;
+import xyz.tomorrowlearncamp.bookking.domain.user.dto.request.UpdateUserRequest;
+import xyz.tomorrowlearncamp.bookking.domain.user.dto.request.UpdateUserRoleRequest;
+import xyz.tomorrowlearncamp.bookking.domain.user.dto.response.UserResponse;
 import xyz.tomorrowlearncamp.bookking.domain.user.entity.User;
 import xyz.tomorrowlearncamp.bookking.domain.user.enums.UserRole;
 import xyz.tomorrowlearncamp.bookking.domain.user.repository.UserRepository;
